Move Sidebar component to options page
