+++ src/options/toggleSidebar.tsx
